Samarth Saarthi Updated Ui/src/functions/requirements.txt
firebase-functions
firebase-admin
fastapi
uvicorn
python-dotenv
google-genai
